+++ app/src/main/java/com/techcourse/controller/HomeController.java
@@ -1,0 +1,17 @@
+package com.techcourse.controller;
+
+import com.interface21.context.stereotype.Controller;
+import com.interface21.web.bind.annotation.RequestMapping;
+import com.interface21.web.bind.annotation.RequestMethod;
+import com.interface21.webmvc.servlet.ModelAndView;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
+@Controller
+public class HomeController {
+
+    @RequestMapping(value = "/", method = RequestMethod.GET)
+    public ModelAndView doGet(final HttpServletRequest request, final HttpServletResponse response) {
+        return new ModelAndView("/index.jsp");
+    }
+}
