+++ app/src/main/java/com/techcourse/controller/LoginController.java
@@ -1,20 +1,23 @@
 package com.techcourse.controller;
 
+import com.interface21.context.stereotype.Controller;
+import com.interface21.web.bind.annotation.RequestMapping;
+import com.interface21.web.bind.annotation.RequestMethod;
 import com.interface21.webmvc.servlet.ModelAndView;
 import com.techcourse.domain.User;
 import com.techcourse.repository.InMemoryUserRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class LoginController implements Controller {
+@Controller
+public class LoginController {
 
     private static final Logger log = LoggerFactory.getLogger(LoginController.class);
 
-    @Override
-    public ModelAndView execute(final HttpServletRequest req, final HttpServletResponse res) {
+    @RequestMapping(value = "/login", method = RequestMethod.POST)
+    public ModelAndView doPost(final HttpServletRequest req, final HttpServletResponse res) {
         if (UserSession.isLoggedIn(req.getSession())) {
             return new ModelAndView("redirect:/index.jsp");
         }
@@ -34,5 +37,15 @@ public class LoginController implements Controller {
             return new ModelAndView("redirect:/index.jsp");
         }
         return new ModelAndView("redirect:/401.jsp");
+    }
+
+    @RequestMapping(value = "/login", method = RequestMethod.GET)
+    public ModelAndView doGet(final HttpServletRequest req, final HttpServletResponse res) {
+        return UserSession.getUserFrom(req.getSession())
+                .map(user -> {
+                    log.info("logged in {}", user.getAccount());
+                    return new ModelAndView("redirect:/index.jsp");
+                })
+                .orElse(new ModelAndView("/login.jsp"));
     }
 }
