--- app/src/main/java/com/techcourse/controller/LoginViewController.java
@@ -1,23 +1,0 @@
-package com.techcourse.controller;
-
-import com.interface21.webmvc.servlet.ModelAndView;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
-
-public class LoginViewController implements Controller {
-
-    private static final Logger log = LoggerFactory.getLogger(LoginViewController.class);
-
-    @Override
-    public ModelAndView execute(final HttpServletRequest req, final HttpServletResponse res) {
-        return UserSession.getUserFrom(req.getSession())
-                .map(user -> {
-                    log.info("logged in {}", user.getAccount());
-                    return new ModelAndView("redirect:/index.jsp");
-                })
-                .orElse(new ModelAndView("/login.jsp"));
-    }
-}
