--- mvc/src/main/java/com/interface21/webmvc/servlet/ControllerHandlerAdapter.java
@@ -1,22 +1,0 @@
-package com.interface21.webmvc.servlet;
-
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-public class ControllerHandlerAdapter implements HandlerAdapter {
-
-    @Override
-    public boolean supports(final Object handler) {
-        return handler instanceof Controller;
-    }
-
-    @Override
-    public ModelAndView handle(
-            final HttpServletRequest request,
-            final HttpServletResponse response,
-            final Object handler
-    ) throws Exception {
-        return ((Controller) handler).execute(request, response);
-    }
-}
